+++ source/Examples/Resources/CMHeartbeatDiscovery/CMHeartbeatDiscovery_Enabled.ps1
@@ -1,0 +1,19 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to enable heartbeat discovery in Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMHeartbeatDiscovery ExampleSettings
+        {
+            SiteCode         = 'Lab'
+            Enabled          = $true
+            ScheduleInterval = 'Days'
+            ScheduleCount    = '4'
+        }
+    }
+}