+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMFallbackStatusPoint Resource
 - Added CMSoftwareUpdatePoint Resource
 - Added CMDsitributionPoint Resource
+- Added CMHeartbeatDiscovery module
+- Added ConvertTo-ScheduleInterval to ResourceHelper
 
 ### Changed